==> Lessons/10_GLMs.Rmd <==
---
title: "10: Generalized Linear Models (T-Test)"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2020"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Describe the components of the generalized linear model (GLM)
2. Apply special cases of the GLM (t-test) to real datasets
3. Interpret and report the results of t-tests in publication-style formats

## Set up
```{r, message = FALSE}
getwd()
library(tidyverse)

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv")

# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

## Generalized Linear Models (GLMs)

The one-sample test (model of the mean), two-sample t-test, analysis of variance (ANOVA), and linear regression are all special cases of the **generalized linear model** (GLM). The GLM also includes analyses not covered in this class, including logistic regression, multinomial regression, chi square, and log-linear models. The common characteristic of general linear models is the expression of a continuous response variable as a linear combination of the effects of categorical or continuous explanatory variables, plus an error term that expresses the random error associated with the coefficients of all explanatory variables. The explanatory variables comprise the deterministic component of the model, and the error term comprises the stochastic component of the model. Historically, artificial distinctions were made between linear models that contained categorical and continuous explanatory variables, but this distinction is no longer made. The inclusion of these models within the umbrella of the GLM allows models to fit the main effects of both categorical and continuous explanatory variables as well as their interactions. 

### Choosing a model from your data: A "cheat sheet"

**T-test:** Continuous response, one categorical explanatory variable with two categories (or comparison to a single value if a one-sample test).

**One-way ANOVA (Analysis of Variance):** Continuous response, one categorical explanatory variable with more than two categories.

**Two-way ANOVA (Analysis of Variance)** Continuous response, two categorical explanatory variables.

**Single Linear Regression** Continuous response, one continuous explanatory variable.

**Multiple Linear Regression** Continuous response, two or more continuous explanatory variables.

**ANCOVA (Analysis of Covariance)** Continuous response, categorical explanatory variable(s) and  continuous explanatory variable(s).

If multiple explanatory variables are chosen, they may be analyzed with respect to their **main effects** on the model (i.e., their separate impacts on the variance explained) or with respsect to their **interaction effects,** the effect of interacting explanatory variables on the model. 

### Assumptions of the GLM

The GLM is based on the assumption that the data residuals approximate a normal distribution (or a linearly transformed normal distribution). We will discuss the non-parametric analogues to several of these tests if the assumptions of normality are violated. For tests that analyze categorical explanatory variables, the assumption is that the variance in the response variable is equal among groups. Note: environmental data often violate the assumptions of normality and equal variance, and we will often proceed with a GLM even if these assumptions are violated. In this situation, justifying the decision to proceed with a linear model must be made. 

## T-Test
### One-sample t-test
The object of a one sample test is to test the null hypothesis that the mean of the group is equal to a specific value. For example, we might ask ourselves (from the EPA air quality processed dataset): 

Are Ozone levels below the threshold for "good" AQI index (0-50)?

```{r}

summary(EPAair$Ozone)
EPAair.subsample <- sample_n(EPAair, 5000) # shapiro test can only take a maximum of 5000 sampples, so we create a subsample

# Evaluate assumption of normal distribution
shapiro.test((EPAair.subsample$Ozone)) #null: data are normally distributed, alt: data are not normally distributed, in this case our data is not nomrally distributed
ggplot(EPAair, aes(x = Ozone)) +
  geom_histogram() #right skewed, postive skew??
qqnorm(EPAair$Ozone); qqline(EPAair$Ozone) ###samples are significantly higher than what we would expect, we just did this we ttest we wont do with other.

O3.onesample <- t.test(EPAair$Ozone, mu = 50, alternative = "less") #### alternative hypothesis is that mean is less than 50,if you do not determine alternative is less, it will determine both less and high. But we want that our ozone shd be less,to be good air quality value. without specifying alternative = less, it would just check if it is equal to 50, i tried it.
O3.onesample # t statistic allows us to get a p-value

Ozone.plot <- ggplot(EPAair, aes(x = Ozone)) + #geom density to plot only needs 1 variable, density distribution of all those values.
  geom_density(stat = "count", fill = "gray") + #showing actual count of data, you can also specify 
  #geom_density(fill = "gray") + #shows density of data, what proportion of data is at or below that point?
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.5) +
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) # this last part of code specifies that ggplot axes coincide at zero.
print(Ozone.plot)
```

Write a sentence or two about the results of this test. Include both the results of the test and an interpretation that puts the findings in context of the resarch question.

> Shapiro wilk test rejected the null hypothesis and therfore informed us that, with a confidence of 95%, the distribution of data are not normally distributed. The t-test informs that, with a 95% conidence interval, the mean average ozone concentration is less than 50.
Now this is what Kateri said, 
EPA ozone measurements for 2018-2019 were significantly lower than 50, the AQI threshold for "good" air quality (one sample t-test; t = -57.98, p < 0.0001), put your pvalue if it is between 0.01 or greater.. she didn't put anything about alternative, null etc, just very concise. You can talk about assumptions in method. THis assumption was violated becaus of xyz. 
one sample t-test you compare one group to a specific value.

### Two-sample t-test
The two-sample *t* test is used to test the hypothesis that the mean of two samples is equivalent. Unlike the one-sample tests, a two-sample test requires a second assumption that the variance of the two groups is equivalent. Are Ozone levels different between 2018 and 2019?
We use this if you want to test about ozone measurements of two different places or two different places 

```{r}
shapiro.test(EPAair$Ozone[EPAair$Year == 2018]) # all the ozone measurements where year is 2018
shapiro.test(EPAair$Ozone[EPAair$Year == 2019]) # all the ozone measurements where the year is 2019
var.test(EPAair$Ozone ~ EPAair$Year) # to test if variance is similar between the two coz assumption is that variance are equal.. #our varinces are different than one another. We violated both the normality and equal variance assumption

ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) + # she did as.factor, you can use this for your exploratory analysis.
  geom_freqpoly()

# Format as a t-test
O3.twosample <- t.test(EPAair$Ozone ~ EPAair$Year) # for two sample t-test instead of comparing with a mu, just give the data regarding the other group
O3.twosample # we rejected null that mean is zero, we can see there is significant increase in ozone concnetation from 2018 to 2019, df effective sample size, you have statistical robustness, it gives us actual increase in mean
O3.twosample$p.value 

# Format as a GLM
O3.twosample2 <- lm(EPAair$Ozone ~ EPAair$Year) # doing same model as 
summary(O3.twosample2) #for each  increase in year  we exoect an increase of 0.84 aqui value, it gives use coefficient with each increment in year

plot(O3.twosample2)
# first thing it shows is residual vs fitted, next it gives us qqplot with residuals, next is scale location you are looking for the red line to be flat, cooks distance indicates that you have a statisitcal outlier
#### now to show all the 4 graphs at same time
par(mfrow = c(2,2)) #2 rows 2 columns
par(mfrow = c(1,1))
```

### Non-parametric equivalent of t-test: Wilcoxon test

When we wish to avoid the assumption of normality, we can apply *distribution-free*, or non-parametric, methods in the form of the Wilcoxon rank sum (Mann-Whitney) test. The Wilcoxon test replaces the data by their rank and calculates the sum of the ranks for each group. Notice that the output of the Wilcoxon test is more limited than its parametric equivalent.
Non-parametric test is distribution free? yeah i guess....
```{r}
O3.onesample.wilcox <- wilcox.test(EPAair$Ozone, mu = 50, alternative = "less")
O3.onesample.wilcox #this gives us less statsitical robustness, like no degrees of freedom...
O3.twosample.wilcox <- wilcox.test(EPAair$Ozone ~ EPAair$Year)
O3.twosample.wilcox #they both give same results, both years have different air quality, but we couldn't know which one is more or less.
```

### Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year . Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
Ozone.plot2 <- ggplot(EPAair, aes(x = Ozone, fill = as.factor(EPAair$Year))) + 
  geom_density(alpha = 0.5, color = NA) + #color = NA removed the black outlines
  #geom_density(fill = "gray") + 
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) 
print(Ozone.plot)

Ozone.boxplot2 <- ggplot(EPAair, aes(y = Ozone, x = as.factor(EPAair$Year), color = as.factor(EPAair$Year))) + 
  geom_boxplot(alpha = 0.5) #color = NA removed the black outlines
  #geom_density(fill = "gray") + 
  #scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) 
print(Ozone.boxplot2)

Ozone.violinplot2 <- ggplot(EPAair, aes(y = Ozone, x = as.factor(EPAair$Year), color = as.factor(EPAair$Year))) + 
  geom_violin(alpha = 0.5) + geom_hline (yintercept = 50) #color = NA removed the black outlines
  #geom_density(fill = "gray") + 
  #scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) 
print(Ozone.violinplot2)
```

Now, write a summary of your findings, incorporating statistical output, reference to the figure(s), and a contextual interpretation.

>
